# api-spec.yaml
openapi: 3.0.3
info:
  title: Riskhub Environment Stock Exchange API
  description: |-
    API Endpoints for the Riskhub Stock Exchange API
    https://github.com/riskhubdev/env-stock-exchange
  version: 1.0.0
servers:
  - url: 'http://localhost:3000'
    description: Development server
...
paths:
...
  '/api/environment/reserve/{environment}':
    post:
      tags:
        - Environment
      parameters:
        - $ref: '#/components/parameters/Environment'
      responses:
        '204':
          description: Environment reserved
        '409':
          description: Environment already reserved
          schema:
            type: object
            required:
              - message
            properties:
              owner:
                type: string
                description: Owner of the environment
                example: 'Vittorio'
        '404':
          description: Environment does not exist
  '/api/environment/release/{environment}':
    post:
      tags:
        - Environment
      parameters:
        - $ref: '#/components/parameters/Environment'
      responses:
        '204':
          description: Environment reserved
        '404':
          description: Environment does not exist
components:
  parameters:
    Environment:
      in: path
      name: environment
      required: true
      schema:
        type: string
      description: Environment
      example: dev01
...